Move make_directories into deploy. Small bug fixes to get it working. NFS mount as nfs, not nfs4
